9-times_table.c

+++ 0x02-functions_nested_loops/9-times_table.c
@@ -7,55 +7,55 @@
  */
 void times_table(void)
 {
-  float a;
+  int a;
 
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%0.2f, ", a*0);
+			printf("%d, ", a*0);
 		}	
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a);
+			printf("%d, ", a);
 		}	
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*2);
+			printf("%d, ", a*2);
 		}
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*3);
+			printf("%d, ", a*3);
 		}
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*4);
+			printf("%d, ", a*4);
 		}
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*5);
+			printf("%d, ", a*5);
 		}
 		printf("\n");
 		for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*6);
+			printf("%d, ", a*6);
 		}
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*7);
+			printf("%d, ", a*7);
 		}
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*8);
+			printf("%d, ", a*8);
 		}
 		printf("\n");
 	for (a = 0; a <= 9; a++)
 		{
-			printf("%f, ", a*9);
+			printf("%d, ", a*9);
 		}	
 }
